server: delete the container if it cannot be restored
If a pod or a container could not be restored, then delete them
from the storage.

Since the container was present in the storage and CRI-O didn't report
it to the the Kubelet, we would have failed the request from the
Kubelet to re-create it over and over again.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=1677509

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>
(cherry picked from commit 67639b573b059fa55635794bcacb557b4909b327)